voice-backend/requirements.txt
fastapi
uvicorn
praat-parselmouth
faster-whisper
numpy
python-multipart
